TextMining-main/requirements.txt
# Core LangChain
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# LLM Providers
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
openai>=1.10.0
langchain-google-genai
google-generativeai

# Hugging Face
huggingface-hub>=0.20.0
transformers>=4.36.0

# Vector store and embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.3.1

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Web interface
streamlit>=1.30.0

# Evaluation (for RAGAS later)
ragas>=0.1.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
